requirements.txt
pandas
openpyxl
pydantic
openai
fastapi
uvicorn
sentence-transformers
chromadb
python-dotenv
passlib[bcrypt]
bcrypt
python-jose[cryptography]
tavily-python
